Ups falto un migration

+++ WorkSynergy.Core.Application/Features/JobOffers/Commands/ChangeStatusJobOffer/ChangeStatusJobOfferCommand.cs
@@ -87,7 +87,7 @@ namespace WorkSynergy.Core.Application.Features.JobOffers.Commands.CreateJobOffe
                         contract.FixedPriceMilestones = new List<FixedPriceMilestone>() { fixedMilestone };
                         break;
                     case (int)ContractOptions.PerHour:
-                        contract.TotalPayment = jobOffer.HourlyRate * jobOffer.TotalHours;
+                        contract.TotalPayment = (double)jobOffer.HourlyRate * jobOffer.TotalHours;
                         HourlyMilestone hourlyMilestone = new()
                         {
                             TotalHours = jobOffer.TotalHours,
